Now noiseless init of DF tensor. AC params are based on error threshold. Lambda-based reward tested. Now plotfilename can be specified directly as cmdline argument to the script (to avoid long ugly names for publication inteded plots). Changed BG integration stopping threshold to allow longer integration to avoid uncomplete convergence.

+++ BG_model.cpp
@@ -341,8 +341,8 @@ void BG_model::init(parmap & params,Exporter *exporter_)
     Q = stof(params["Q"]);
     A_exp = stof(params["A_exp"]);
 
-    na = stof(params["na"]);
-    nc = stof(params["nc"]);
+    na = stoi(params["na"]);
+    nc = stoi(params["nc"]);
 
     HD_D2activity_reduction = 50.;
     PD_LTP_reduction = 80.;
